+++ CHANGELOG.md
@@ -12,6 +12,11 @@ Global Functions
 
 - `FIXME(_:dsohandle:file:line:function:)` method is added
 
+Macros
+
+- Macros are added to `VCore`
+- `URL(_:)` macro is added, that creates a non-optional `URL` from a `String` literal
+
 ### [5.2.1(88)](https://github.com/VakhoKontridze/VCore/releases/tag/5.2.1) — *2023 12 08*
 
 Other